Extract NonceVerifier
