+++ packages/processor/source/verifiers/index.ts
@@ -1,2 +1,3 @@
 export * from "./incompatible-transactions-verifier";
+export * from "./nonce-verifier";
 export * from "./verify-block-verifier";
